reorganize test directory because claude was over-eager
