+++ test/lua/helpers/create_contract.test.cpp
@@ -1,0 +1,43 @@
+#include "modules/lua/helpers.h"
+#include "modules/rocket/rocket_module.h"
+#include <catch2/catch_test_macros.hpp>
+#include <flecs.h>
+
+SCENARIO("create_contract", "[helpers][lua]") {
+  flecs::world world;
+  world.import<RocketModule>();
+  world.entity("Contracts");
+
+  GIVEN("a unique contract name") {
+    WHEN("create_contract is called") {
+      auto contract = create_contract(world, "Mission1", "AcmeCorp",
+                                      "Deliver satellite", 2000.0f, 8000.0f);
+      THEN("a valid entity is returned") { REQUIRE(contract.is_valid()); }
+      THEN("the Contract component has the correct fields") {
+        auto &c = contract.get<Contract>();
+        CHECK(c.client == "AcmeCorp");
+        CHECK(c.description == "Deliver satellite");
+        CHECK(c.upfront_payment == 2000.0f);
+        CHECK(c.completion_payment == 8000.0f);
+      }
+      THEN("the entity is a child of the Contracts node") {
+        CHECK(contract.parent() == world.lookup("Contracts"));
+      }
+    }
+  }
+
+  GIVEN("a name that already exists under Contracts") {
+    auto existing = world.entity("Dup")
+                        .child_of(world.lookup("Contracts"))
+                        .set<Contract>({.client = "OldClient",
+                                        .description = "OldDesc",
+                                        .upfront_payment = 0,
+                                        .completion_payment = 0,
+                                        .status = ContractStatus::Open,
+                                        .failed = false});
+    WHEN("create_contract is called with the same name") {
+      auto contract = create_contract(world, "Dup", "NewClient", "NewDesc", 0, 0);
+      THEN("the existing entity is returned") { CHECK(contract == existing); }
+    }
+  }
+}
